rabbit: Run plugins' boot steps during rabbit start/2
This restores the behavior prior the commit making `rabbit` closer to a
standard Erlang application.

Plugins are still actually started after rabbit is started (because they
depend on the `rabbit` application). Only the execution of their boot
steps was moved earlier.

With the behavior restored, it also means that a plugin's dependencies
are not started yet when its boot steps are executed.